+++ __tests__/screens/home.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { shallow, mount } from 'enzyme';
+import { shallow } from 'enzyme';
 import Home from '../../src/screens/Home';
 
 describe('Home screen', () => {
