+++ src/screens/Home.js
@@ -25,20 +25,58 @@ export default class Home extends Component {
     super(props);
     this.state = {
       engineers: [],
-      loading: true
+      page: 1,
+      loading: true,
+      loadingMoreData: false
     };
   }
 
   componentDidMount = () => {
     this.setState({ loading: true });
+    this.makeRemoteRequest();
+  };
+
+  makeRemoteRequest = () => {
+    const { page } = this.state;
     return fetch(
-      'https://api.github.com/search/users?q=location:lagos+language:java'
+      `https://api.github.com/search/users?q=location:lagos+language:java&per_page=20&page=${page}`
     )
       .then(response => response.json())
       .then(response => {
-        this.setState({ engineers: response.items, loading: false });
+        this.setState({
+          engineers: [...this.state.engineers, ...response.items],
+          loading: false,
+          loadingMoreData: false
+        });
       })
       .catch(error => this.setState({ error: error }));
+  };
+
+  handleLoadMore = () => {
+    this.setState(
+      {
+        page: this.state.page + 1,
+        loadingMoreData: true
+      },
+      () => {
+        this.makeRemoteRequest();
+      }
+    );
+  };
+
+  renderFooter = () => {
+    const { loadingMoreData } = this.state;
+    return loadingMoreData ? (
+      <View
+        style={{
+          paddingVertical: 20,
+          borderTopWidth: 1,
+          borderTopColor: '#0000ff'
+        }}
+      >
+        <ActivityIndicator size="large" color="#0000ff" />
+      </View>
+    ) : null;
   };
   render() {
     const { engineers, loading } = this.state;
@@ -50,7 +88,10 @@ export default class Home extends Component {
         <FlatList
           data={[...engineers]}
           renderItem={({ item }) => <UserItem profile={item} />}
-          keyExtractor={(item, index) => `${item.id}`}
+          keyExtractor={(item, index) => `${item.id + index}`}
+          ListFooterComponent={this.renderFooter}
+          onEndReached={this.handleLoadMore}
+          onEndReachedThreshold={1}
         />
       </View>
     );
